requirements.txt
python-dotenv
discord
colorama
requests
aiomysql
aiofiles
pillow
easy_pil